+++ index.css
@@ -3,7 +3,6 @@
   background-size: "cover";
   background-attachment: fixed;
   font-family: Inter;
-  height: 100vh;
 }
 
 .datePicker {
